Using same navigator for android and iOS Moved save/delete to nav bar
