+++ app/js/index.android.js
@@ -28,98 +28,8 @@
 var React = require('react-native');
 var {
     AppRegistry,
-    Image,
-    Navigator,
-    StyleSheet,
-    Text,
-    TouchableNativeFeedback,
-    View,
 } = React;
 
-var storeMgr = require('./StoreMgr');
-var SearchScreen = require('./SearchScreen');
-var ContactScreen = require('./ContactScreen');
-
-var NavigationBarRouteMapper = {
-    LeftButton: function(route, navigator, index, navState) {
-        if (route.name === "Contact") {
-            return (<TouchableNativeFeedback onPress={() => navigator.pop()}>
-                      <View style={styles.navBarElt}>
-                        <Image source={require('image!android_back_white')} style={styles.icon}/>
-                      </View>
-                    </TouchableNativeFeedback>);
-        }
-    },
-
-    RightButton: function(route, navigator, index, navState) {
-        if (route.name === "Contacts") {
-            return (<TouchableNativeFeedback onPress={() => storeMgr.reSyncData()}>
-                      <View style={styles.navBarElt}>
-                        <Image source={require('image!sync')} style={styles.icon}/>
-                      </View>
-                    </TouchableNativeFeedback>);
-        }
-    },
-
-    Title: function(route, navigator, index, navState) {
-        return ( <View style={styles.navBarElt}><Text style={styles.navBarText}> {route.name} </Text></View>);
-  },
-
-};
-
-var App = React.createClass({
-    componentDidMount: function() {
-        storeMgr.syncData();
-    },
-
-
-    render: function() {
-        var initialRoute = {name: 'Contacts'};
-        return (
-                <Navigator
-                  style={styles.container}
-                  initialRoute={initialRoute}
-                  configureScene={() => Navigator.SceneConfigs.PushFromRight}
-                  renderScene={(route, navigator) => {
-                      if (route.name === 'Contacts') {
-                          return (<SearchScreen style={styles.scene} navigator={navigator} />);
-                      }
-                      else if (route.name === 'Contact') {
-                          return (<ContactScreen style={styles.scene} navigator={navigator} contact={route.contact} />);
-                      }
-                  }}
-                  navigationBar={<Navigator.NavigationBar routeMapper={NavigationBarRouteMapper} style={styles.navBar} />} />
-        );
-    }
-});
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'white',
-  },
-  navBar: {
-    backgroundColor: 'red',
-    height: 56,
-  },
-  navBarElt: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems:'center',
-  },
-  navBarText: {
-    fontSize: 20,
-    color: 'white',
-  },
-  icon: {
-    width: 24,
-    height: 24,
-  },
-  scene: {
-    flex: 1,
-    paddingTop: 56,
-    backgroundColor: '#EAEAEA',
-  },
-});
+var App = require('./App');
 
 AppRegistry.registerComponent('SmartSyncExplorerReactNative', () => App);
